feat: #1917 Restructure webhook folders
