--- packages/developer-portal/src/components/pages/developer-webhooks/index.ts
@@ -1,2 +1,0 @@
-import DeveloperWebhooks from './developer-webhooks'
-export default DeveloperWebhooks
