+++ packages/developer-portal/src/components/pages/webhooks/__test__/webhooks.test.tsx
@@ -11,7 +11,7 @@ import {
   MODAL_TYPE,
   openCreateModal,
   openTestModal,
-} from '../developer-webhooks'
+} from '../webhooks'
 import { mount } from 'enzyme'
 import { TopicModel } from '@/reducers/webhook-subscriptions'
