+++ packages/developer-portal/src/components/pages/webhooks/index.ts
@@ -1,0 +1,2 @@
+import DeveloperWebhooks from './webhooks'
+export default DeveloperWebhooks
